requirements.txt
python-Levenshtein
jellyfish
scikit-learn
networkx
